Clean up constants folder and refine e2e tests
