+++ README.md
@@ -31,6 +31,31 @@ npm install
 npm start
 ```
 
+### Project structure
+
+The project is split into the following parts:
+
+- `cypress` holds logic for the Cypress end-to-end testing suite tool.
+- `/server` holds some miscellaneous scripts and database migration stuff for use by the backend.
+- `/src` contains most of the code:
+  - `/api` contains client code for interacting with the graphql api.
+  - `/assets` contains images, fonts, and other static assets.
+  - `/components` contains all components that are shared between different parts of the project.
+  - `/context` contains the global state management system.
+  - `/pages` is relatively self-explanatory: it holds the pages of the app.
+  - `/shared` contains all shared code like constants, etc that are used in _multiple_ parts of the project.
+  - `/theme` contains all code dealing with styles and theming of the app.
+
+For more information, consult the README in the root of each directory.
+
+#### GraphQL
+
+All code dealing with GraphQL resides in `/graphql` directories. These directories are co-located next to the parts of the applications using them.
+
+Each of these will contain logic like queries, fragments, etc. The types associated with the logic (generated by Apollo tooling) reside in a `/graphql/types` subdirectory. The mocks for the logic (used in testing and local dev) reside in a `/graphql/mocks` subdirectory.
+
+### Development conventions
+
 This project (loosely) follows a [trunk based development](https://trunkbaseddevelopment.com/) style.
 
 - Branches are split off from the `master` branch for features, fixes, and all other development.
