+++ src/api/graphql/mocks/index.ts
@@ -1,0 +1,1 @@
+export default 1;
