+++ src/api/graphql/mocks/schema.gql
@@ -1,0 +1,85 @@
+enum SalaryPeriod {
+  HOURLY
+  WEEKLY
+  MONTHLY
+}
+
+type Company {
+  id: ID!
+  updatedAt: String!
+
+  name: String!
+  slug: String!
+  desc: String
+
+  logoImg: File
+  logoColor: String
+
+  numRatings: Int!
+  totRating: Int!
+  avgRating: Float!
+  totLearningMentorshipRating: Int!
+  avgLearningMentorshipRating: Float!
+  totMeaningfulWorkRating: Int!
+  avgMeaningfulWorkRating: Float!
+  totWorkLifeBalanceRating: Int!
+  avgWorkLifeBalanceRating: Float!
+
+  jobs: [Job!]
+  reviews: [Review!]
+}
+
+type Job {
+  id: ID!
+  updatedAt: String!
+
+  name: String!
+  slug: String!
+  location: String
+
+  minHourlySalary: Int!
+  maxHourlySalary: Int!
+  avgHourlySalary: Int!
+  hourlySalaryCurrency: String!
+
+  numRatings: Int!
+  totRating: Int!
+  avgRating: Float!
+  totLearningMentorshipRating: Int!
+  avgLearningMentorshipRating: Float!
+  totMeaningfulWorkRating: Int!
+  avgMeaningfulWorkRating: Float!
+  totWorkLifeBalanceRating: Int!
+  avgWorkLifeBalanceRating: Float!
+
+  company: [Company!]
+  reviews: [Review!]
+}
+
+type Review {
+  id: ID!
+  updatedAt: String!
+
+  author: String!
+  body: String!
+  tags: String
+  salary: Int!
+  salaryCurrency: String!
+  salaryPeriod: SalaryPeriod!
+
+  overallRating: Float!
+  learningMentorshipRating: Float!
+  meaningfulWorkRating: Float!
+  workLifeBalanceRating: Float!
+
+  isLegacy: Boolean!
+  isAnonymous: Boolean!
+  legacyUpdatedAt: String
+
+  job: Job!
+  company: Company!
+}
+
+type File {
+  downloadUrl: String!
+}
