+++ src/api/graphql/types/CompanyResult.ts
