+++ src/api/graphql/types/JobResult.ts
