+++ src/pages/companies/graphql/queries.ts
@@ -1,5 +1,5 @@
 import { gql } from "apollo-boost";
-import { jobResultFragment } from "src/api/fragments";
+import { jobResultFragment } from "src/api/graphql";
 
 export const GET_COMPANY_DETAILS = gql`
   query GetCompanyDetails($slug: String) {
