+++ src/pages/jobs/components/JobsPage.tsx
@@ -12,7 +12,7 @@ import {
 } from "src/shared/hooks/useSearch";
 import { RESULTS_PER_PAGE } from "src/shared/constants/search";
 
-import { IReviewUserCardItem } from "src/shared/types";
+import { IReviewUserCardItem } from "src/shared/constants/card";
 import { GetJobDetails } from "../graphql/types/GetJobDetails";
 import { GetJobReviews } from "../graphql/types/GetJobReviews";
 import { GET_JOB_DETAILS, GET_JOB_REVIEWS } from "../graphql/queries";
