+++ src/pages/jobs/graphql/utils.ts
@@ -2,7 +2,8 @@
 import TimeAgo from "javascript-time-ago";
 import en from "javascript-time-ago/locale/en";
 
-import { IJobDetails, IReviewUserCardItem } from "src/shared/types";
+import { IReviewUserCardItem } from "src/shared/constants/card";
+import { IJobDetails } from "../components/JobDetailsCard";
 import { GetJobDetails_job } from "./types/GetJobDetails";
 import {
   GetJobReviews,
