+++ src/shared/README.md
@@ -1,3 +1,5 @@
 ## Shared code
 
 Any reusable logic or function in the app such as custom hooks or analytics is defined here.
+
+All reusable and data schema-agnostic types are defined here as well.
