+++ src/shared/constants/card.ts
