+++ src/shared/hooks/useSearch.ts
@@ -11,7 +11,7 @@ import { useState, useCallback, useEffect, useMemo } from "react";
 import { useSearchParams } from "src/shared/hooks/useSearchParams";
 import { RESULTS_PER_PAGE } from "src/shared/constants/search";
 import { analytics } from "src/shared/utils/analytics";
-import { IGenericCardItem } from "src/shared/types";
+import { IGenericCardItem } from "src/shared/constants/card";
 
 /*******************************************************************
  *                             **Types**                           *
