--- src/shared/types/README.md
@@ -1,3 +1,0 @@
-## Types
-
-All reusable and data schema-agnostic types are defined here.
