--- src/shared/types/index.ts
@@ -1,2 +1,0 @@
-export * from "./card";
-export * from "./details";
